// src/types/enum.ts
export const enum ActionLoad {
  GET_LIST_BOOKS = 'GET_BOOKS',
  GET_CATEGORY_BOOKS = 'GET_CATEGORY_BOOKS',
  GET_ID_BOOK = 'GET_ID_BOOK',
  SET_PATH = 'SET_PATH',
  AUTH = 'AUTH',
  BODY_AUTH = 'BODY_AUTH',
  SET_DATE_ORDER = 'SET_DATE_ORDER',
  SET_BOOK = 'SET_BOOK',
  LOADED = 'TOGGLE_LOADED',
  ERROR = 'TOGGLE_ERROR',
  CLOSE_WINDOW = 'CLOSE_WINDOW',
  SET_DATA_BOOKING = 'SET_DATA_BOOKING',
  ERROR_BOOKING = 'ERROR_BOOKING',
  SET_COMMENT = 'SET_COMMENT',
}

export const enum FormEditFielfs {
  EMAIL = 'email',
  LOGIN = 'login',
  PASSWORD = 'password',
  FIRSTNAME = 'firstName',
  LASTNAME = 'lastName',
  PHONE = 'phone',
}

export const enum PlugFlag {
  NULL = 'null',
  DELAY = 'delay',
}